app/requirements.txt
fastapi~=0.115.6
uvicorn~=0.34.0
httpx~=0.28.1
sqlalchemy~=2.0.36
google-cloud-secret-manager~=2.22.0
pydantic-settings~=2.2.1
google-cloud-storage~=2.19.0
cloud-sql-python-connector[pymysql]~=1.15.0
pymysql~=1.1.1
pandas~=2.2.3
gcsfs~=2024.10.0

pytest~=8.3.4


